requirements.txt
torch==2.7.1
torchvision==0.22.1
matplotlib==3.10.3
pillow==11.3.0